linux: add in-tree device tree overlay support
Allow to deploy in-tree device tree overlays. This implementation
assumes the overlays get built by the given kernel configuration and
are available at the overlays/ directory after the build. This is
useful to deploy the device tree overlays of the Raspberry Pi kernel.

Signed-off-by: Stefan Agner <stefan@agner.ch>

Squashed with the following commits from 2024.02.x-haos:

* linux: fix add in-tree device tree overlay support
* Install overlay_map.dtb when using in-tree DT overlays

Signed-off-by: Jan Čermák <sairon@sairon.cz>