linux-firmware: add option for firmware files compression
Add option for compression of the firmware files. This can save some
space in the rootfs if less efficient (or none) compression is used for
the whole FS. The files are compressed in the install stage, because
different behavior is needed when being installed to the target or to
the images directory (where there are later picked up by the kernel
build). This means that the target directory may contain both
compressed and uncompressed images if the option is changed between
builds without cleaning in between, but similar inconsistencies may be
encountered when selecting individual firmwares to deploy.

Signed-off-by: Jan Čermák <sairon@sairon.cz>