package/go: bump to version 1.25.1
For relase notes, see:
https://go.dev/doc/devel/release#go1.25.0

Building Go 1.24 and later requires Go 1.22.6 or later for bootstrap.

To support this we use our previous Go version 1.23.12 as the version
for go-bootstrap-stage4 and have the build for Go 1.25.1 depend on
go-bootstrap-stage4.

Go version 1.23.12 is the latest go version we can build using
go-bootstrap-stage3.

The package build for go-bootstrap-stage4 if effectively identical to
go-bootstrap-stage3 with only the Go version and stage number changed.

Note: go version 1.25.1 release note mention a security fix
which is for CVE-2025-47910. This commit is not marked as security
bump, because the issue was introduced in version 1.25.0 which
was not included in Buildroot. See:
https://pkg.go.dev/vuln/GO-2025-3955

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
Tested-by: Christian Stewart <christian@aperture.us>
[Julien: add link to release notes and
  comment about CVE in commit log.
]
Signed-off-by: Julien Olivain <ju.o@free.fr>
(cherry picked from commit ecf12c186f1e62751980eb4c7345decd65305680)
Signed-off-by: Jan Čermák <sairon@sairon.cz>